Get rid of ForecastViewC, implement forecastTableView on CurrentLVC
